Add contact management only by owner

+++ src/services/contacts.js
@@ -2,10 +2,10 @@ import { SORT_ORDER } from "../constants/index.js";
 import { ContactsCollection } from "../db/models/contact.js";
 import { calculatePaginationData } from "../utils/calculatePaginationData.js";
 
-export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_ORDER.ASC, sortBy = "_id", filter = {}}) => {
+export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_ORDER.ASC, sortBy = "_id", filter = {}}, uid) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
-    const contactsQuery = ContactsCollection.find();
+    const contactsQuery = ContactsCollection.find({ userId: uid });
     if(filter.contactType) {
         contactsQuery.where("contactType").equals(filter.contactType);
     }
@@ -21,23 +21,23 @@ export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_O
     };
 };
 
-export const getContactById = async contactId => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, uid) => {
+    const contact = await ContactsCollection.find({ userId: uid }).findById(contactId);
     return contact;
 };
 
-export const createContact = async payload => {
-    const newContact = await ContactsCollection.create(payload);
+export const createContact = async (payload, uid) => {
+    const newContact = await ContactsCollection.create({ ...payload, userId: uid });
     return newContact;
 };
 
-export const deleteContactById = async contactId => {
-    const contact = await ContactsCollection.findByIdAndDelete(contactId);
+export const deleteContactById = async (contactId, uid) => {
+    const contact = await ContactsCollection.find({ userId: uid }).findByIdAndDelete(contactId);
     return contact;
 }
 
-export const updateContact = async (contactId, payload, options = {}) => {
-    const result = await ContactsCollection.findByIdAndUpdate(contactId, payload, {new: true, includeResultMetadata: true, ...options});
+export const updateContact = async (contactId, payload, uid, options = {}) => {
+    const result = await ContactsCollection.find({ userId: uid }).findByIdAndUpdate(contactId, payload, {new: true, includeResultMetadata: true, ...options});
     if(!result || !result.value) {
         return null;
     }
